feat: Remove simplejson + Remove Schema validation
