+++ restful_aws_lambda/__init__.py
@@ -2,6 +2,6 @@
 pylambdarest - RESTful API with AWS Lambda and API Gateway.
 """
 
-from pylambdarest.route import route
+from restful_aws_lambda.route import route
 
 __all__ = ["route"]
