+++ restful_aws_lambda/request.py
@@ -4,10 +4,8 @@ Request
 Information parsed from the API Gateway event.
 
 """
-
+import json
 from typing import Optional
-
-import simplejson as json
 
 
 class Request:
